Refactor and improve is_composite function
Updated the is_composite function for better efficiency and made it more descriptive. Firstly, the parameter name has been changed from 'n' to 'number' for clarity. Secondly, added a docstring to provide a clear definition of what a composite number is. Lastly, the function now checks for numbers greater than 3 as any number less than 4 can not be composite. This reduces unnecessary calculations. This change also reflected in the solution function where the range for the loop starts from 4 instead of 1.