Refactor triangle function in main.py
Streamlined the 'solution' function in the main.py file by refactoring the initialization of the array 'arr' and renaming the function 'triangle' to 'fill_triangle' for better readability. The 'fill_triangle' function was also simplified to rule out unnecessary for-loop condition checks thereby, making it more efficient. The argument condition for recursion in 'fill_triangle' function was also updated for correctness.