Add solution and tests for lesson 72413
Added a Python solution for lesson 72413. The solution employs a Floyd-Warshall algorithm to find the minimum cost between any two points in a directed, weighted graph. This is particularly applicable to graph problems involving shortest paths. Tests were also added to confirm the correctness of the solution, covering different types of fare structures.