Add solution and tests for lesson 12953
Added a new Python solution for lesson 12953 in 'main.py'. This solution finds the least common multiple (LCM) of an array of numbers using the reduce function and a custom LCM function. Also included a Greatest Common Divisor (GCD) helper function that's used in calculating the LCM. Two tests have been added in 'test.py' to ensure the solution works as expected.